error handling

+++ bot.py
@@ -94,8 +94,10 @@ async def on_message(message):
     Only reacts to commands when entered in the bot channel.
     """
     role = discord.utils.get(message.author.roles, id=LINKER_ROLE)
-    member = discord.utils.get(message.author)
-    role = discord.utils.get(member.roles, id=LINKER_ROLE)
+    author = message.author
+    role = discord.utils.get(message.guild.roles, id=LINKER_ROLE)
+    if message.author.id == BOT_ID:
+        return
     for i in badwords:
         if i in message.content:
             await message.delete()
@@ -105,7 +107,7 @@ async def on_message(message):
             bot.dispatch('profanity', message, i)
             break
     if 'https://' in message.content or 'http://' in message.content:
-        if role is None:
+        if role not in author.roles:
             await message.delete()
             e_embed.clear_fields()
             e_embed.add_field(name="Erreur", value=f"Tiens, ton lien du Q.")
